add environment to test build #16813
