+++ projects/web-element/src/app/app.module.ts
@@ -219,7 +219,7 @@ export const httpTranslateLoader = (http: HttpClient) =>
         ({ nativeElement: document.body } as ElementRef),
     },
   ],
-  bootstrap: [AppComponent],
+  // bootstrap: [AppComponent],
 })
 export class AppModule implements DoBootstrap {
   constructor(
